Move AnalyzerPluginInfo to analyzer/analyzer_plugin_info.h
Part of a broader effort to clean up include dependencies as a step
towards a modular build.

Created analyzer/analyzer_plugin_info.h containing only
AnalyzerPluginInfo (a plain data class). The new header depends solely
on <QString>.

Removed the AnalyzerPluginInfo definition from
analyzer/plugins/analyzerplugin.h, replacing it with an include of the
new thin header.

Replaced #include "analyzer/plugins/analyzerplugin.h" with
  - preferences/dialog/dlgprefbeats.h
  - preferences/dialog/dlgprefkey.h

These headers only need the AnalyzerPluginInfo type (for a
QList<AnalyzerPluginInfo> member). They no longer drag in the full
analyzer plugin hierarchy (audio/frame.h, track/beats.h, track/keys.h,
etc.).

Added a direct #include "proto/keys.pb.h" to
preferences/keydetectionsettings.h, which was relying on
mixxx::track::io::key::ChromaticKey being included transitively through
analyzerplugin.h → track/keys.h.
