+++ src/analyzer/plugins/analyzerplugin.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <QString>
-
+#include "analyzer/analyzer_plugin_info.h"
 #include "audio/frame.h"
 #include "track/beats.h"
 #include "track/bpm.h"
@@ -9,41 +8,6 @@
 #include "util/types.h"
 
 namespace mixxx {
-
-class AnalyzerPluginInfo {
-  public:
-    AnalyzerPluginInfo(const QString& id,
-            const QString& author,
-            const QString& name,
-            bool isConstantTempoSupported)
-            : m_id(id),
-              m_author(author),
-              m_name(name),
-              m_isConstantTempoSupported(isConstantTempoSupported) {
-    }
-
-    const QString& id() const {
-        return m_id;
-    }
-
-    const QString& author() const {
-        return m_author;
-    }
-
-    const QString& name() const {
-        return m_name;
-    }
-
-    bool isConstantTempoSupported() const {
-        return m_isConstantTempoSupported;
-    }
-
-  private:
-    QString m_id;
-    QString m_author;
-    QString m_name;
-    bool m_isConstantTempoSupported;
-};
 
 class AnalyzerPlugin {
   public:
